revision v1

+++ app/Http/Controllers/GuestController.php
@@ -3,14 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Guest;
-use Barryvdh\DomPDF\Facade as PDF;
 use Exception;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Http\Response;
 
 class GuestController extends BaseController
 {
@@ -27,42 +25,36 @@ class GuestController extends BaseController
     /**
      * @throws Exception
      */
-    public function data()
+    public function data(): JsonResponse
     {
-        $guest = Guest::orderBy('created_at')->get();
+        try {
+            $guest = Guest::orderBy('created_at')->get();
 
-        return datatables()
-            ->of($guest)
-            ->addIndexColumn()
-            ->addColumn('select_all', function ($guest) {
-                return '
-                    <input type="checkbox" name="guest_id[]" value="'. $guest->guest_id .'">
+            return datatables()
+                ->of($guest)
+                ->addIndexColumn()
+                ->addColumn('select_all', function ($guest) {
+                    return '
+                    <input type="checkbox" name="guest_id[]" value="'. e($guest->guest_id) .'">
                 ';
-            })
-            ->addColumn('action', function ($guest) {
-                return '
-                <div class="btn-group">
-                    <button type="button" onclick="updateOne(`'. route('guest.update', $guest->guest_id) .'`)" class="btn btn-xs btn-info btn-flat">
-                        <i class="fa fa-pencil"> edit</i>
-                    </button>
-                    <button type="button" onclick="deleteOne(`' . route('guest.destroy', $guest->guest_id) . '`, `' . $guest->name . '`)" class="btn btn-xs btn-danger btn-flat">
-                        <i class="fa fa-trash"> delete</i>
-                    </button>
-                </div>
+                })
+                ->addColumn('action', function ($guest) {
+                    return '
+                    <div class="btn-group">
+                        <button type="button" onclick="updateOne(`'. e(route('guest.update', $guest->guest_id)) .'`)" class="btn btn-xs btn-info btn-flat">
+                            <i class="fa fa-pencil"> edit</i>
+                        </button>
+                        <button type="button" onclick="deleteOne(`' . e(route('guest.destroy', $guest->guest_id)) . '`, `' . e($guest->name) . '`)" class="btn btn-xs btn-danger btn-flat">
+                            <i class="fa fa-trash"> delete</i>
+                        </button>
+                    </div>
                 ';
-            })
-            ->rawColumns(['action', 'select_all', 'created_at'])
-            ->make(true);
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return void
-     */
-    public function create()
-    {
-        //
+                })
+                ->rawColumns(['action', 'select_all'])
+                ->make(true);
+        } catch (Exception $e) {
+            return response()->json(['error' => 'Failed to load data'], 500);
+        }
     }
 
     /**
@@ -73,9 +65,16 @@ class GuestController extends BaseController
      */
     public function store(Request $request): JsonResponse
     {
+        $validatedData = $request->validate([
+            'name' => 'required|string|max:255',
+            'phone_number' => 'required|string|max:20',
+            'address' => 'nullable|string|max:500',
+        ]);
+
         $guest = Guest::latest()->first() ?? new Guest();
-        $request['member_code'] = 'M'. add_zero((int)$guest->guest_id +1, 6);
-        return $this->saveModel($request, new Guest());
+        $validatedData['member_code'] = 'M'. add_zero((int)$guest->guest_id + 1, 6);
+
+        return $this->saveModel($validatedData, new Guest());
     }
 
     /**
@@ -91,17 +90,6 @@ class GuestController extends BaseController
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param int $id
-     * @return void
-     */
-    public function edit(int $id)
-    {
-        //
-    }
-
-    /**
      * Update the specified resource in storage.
      *
      * @param Request $request
@@ -110,24 +98,33 @@ class GuestController extends BaseController
      */
     public function update(Request $request, int $id): JsonResponse
     {
+        $validatedData = $request->validate([
+            'name' => 'required|string|max:255',
+            'phone_number' => 'required|string|max:20',
+            'address' => 'nullable|string|max:500',
+        ]);
+
         $guest = Guest::find($id);
         if (!$guest) {
             return response()->json('Error: Guest not found.', 404);
         }
-        return $this->saveModel($request, $guest);
+
+        return $this->saveModel($validatedData, $guest);
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param int $id
-     * @return Response
+     * @return JsonResponse
      */
-    public function destroy(int $id): Response
+    public function destroy(int $id): JsonResponse
     {
         $guest = Guest::find($id);
-        $guest->delete();
-
-        return response(null, 204);
+        if ($guest) {
+            $guest->delete();
+            return response()->json(['message' => 'Operation successful'], 200);
+        }
+        return response()->json(['error' => 'Guest not found'], 404);
     }
 }
